debug

+++ backend/logout/user/logout.php
@@ -13,11 +13,13 @@ $user_id = $row['user_id'];
 
 $sql = "SELECT on_service FROM websocket WHERE id = '$user_id' and is_driver = 0";
 $result = $conn->query($sql);
-$row = $result->fetch_assoc();
-$on_service = $row['on_service'];
-if ($on_service == 0) {
-    $sql = "DELETE FROM booking WHERE user_id = '$user_id'";
-    $conn->query($sql);
+if ($result->num_rows == 1) {
+    $row = $result->fetch_assoc();
+    $on_service = $row['on_service'];
+    if ($on_service == 0) {
+        $sql = "DELETE FROM booking WHERE user_id = '$user_id'";
+        $conn->query($sql);
+    }
 }
 
 $sql2 = "UPDATE `user` SET `status` = 0 WHERE `user_id` = $user_id";
